+++ core-ng/src/main/java/core/framework/api/util/JSON.java
@@ -35,4 +35,12 @@ public final class JSON {
             throw new UncheckedIOException(e);
         }
     }
+
+    public static <T extends Enum> T fromEnumValue(Class<T> valueType, String jsonValue) {
+        return JSONMapper.OBJECT_MAPPER.convertValue(jsonValue, valueType);
+    }
+
+    public static <T extends Enum> String toEnumValue(T value) {
+        return JSONMapper.OBJECT_MAPPER.convertValue(value, String.class);
+    }
 }