+++ core-ng/src/main/java/core/framework/api/web/exception/NotFoundException.java
@@ -8,7 +8,8 @@ import core.framework.api.log.Warning;
  */
 @Warning
 public final class NotFoundException extends RuntimeException implements ErrorCode {
-    private static final String DEFAULT_ERROR_CODE = "NOT_FOUND";
+    public static final String DEFAULT_ERROR_CODE = "NOT_FOUND";
+
     private static final long serialVersionUID = 8663360723004690205L;
 
     private final String errorCode;